backend/requirements.txt
flask==2.2.2
flask-cors==3.0.10
Flask-RESTful==0.3.9
flask-apispec==0.11.4
marshmallow-enum==1.5.1
python-dotenv==0.21.0
pytest==7.2.0
tox==3.27.1
SQLAlchemy==1.4.44
